+++ src/main/java/org/cryptimeleon/craco/sig/StructurePreservingSignatureEQScheme.java
@@ -1,14 +1,8 @@
 package org.cryptimeleon.craco.sig;
 
 
-import org.cryptimeleon.craco.common.plaintexts.GroupElementPlainText;
-import org.cryptimeleon.craco.common.plaintexts.MessageBlock;
 import org.cryptimeleon.craco.common.plaintexts.PlainText;
-import org.cryptimeleon.math.structures.groups.GroupElement;
-import org.cryptimeleon.math.structures.groups.cartesian.GroupElementVector;
 import org.cryptimeleon.math.structures.rings.zn.Zn;
-
-import java.util.function.Function;
 
 /**
  * A structure-preserving signature scheme on equivalence classes (SPS-EQ).
@@ -22,7 +16,7 @@ import java.util.function.Function;
  * and Constant-Size Anonymous Credentials", in Cryptology ePrint Archive, Report
  * 2014/944, 2014.
  */
-public interface StructurePreservingSignatureEQScheme extends StandardMultiMessageSignatureScheme {
+public interface StructurePreservingSignatureEQScheme extends MultiMessageStructurePreservingSignatureScheme {
 
     /**
      * Returns a signature for the new representative computed based on the previous representative and the given
@@ -59,56 +53,4 @@ public interface StructurePreservingSignatureEQScheme extends StandardMultiMessa
      */
     PlainText chgRepMessage(PlainText plainText, Zn.ZnElement mu);
 
-    /**
-     * Signs multiple group elements as a single unit.
-     * @param secretKey key to sign with
-     * @param groupElements group elements to sign
-     * @return signature over the given plaintexts
-     */
-    default Signature sign(SigningKey secretKey, GroupElement... groupElements) {
-        return sign(secretKey, new GroupElementVector(groupElements));
-    }
-
-    /**
-     * Verifies a signature for multiple messages.
-     * @param verificationKey key to use for verification
-     * @param signature signature to verify
-     * @param groupElements group elements to verify signature for
-     * @return true if verification succeeds, else false
-     */
-    default Boolean verify(VerificationKey verificationKey, Signature signature, GroupElement... groupElements) {
-        return verify(verificationKey, signature, new GroupElementVector(groupElements));
-    }
-
-    /**
-     * Signs the given vector of group elements.
-     * @param secretKey key to sign with
-     * @param groupElements vector of group elements to sign
-     * @return signature over the given vector of plaintexts
-     */
-    default Signature sign(SigningKey secretKey, GroupElementVector groupElements) {
-        return sign(
-                new MessageBlock(groupElements.map(
-                        (Function<GroupElement, GroupElementPlainText>) GroupElementPlainText::new
-                )),
-                secretKey
-        );
-    }
-
-    /**
-     * Verifies a signature for a vector of group elements.
-     * @param verificationKey key to use for verification
-     * @param signature signature to verify
-     * @param groupElements vector of group elements to verify signature for
-     * @return true if verification succeeds, else false
-     */
-    default Boolean verify(VerificationKey verificationKey, Signature signature, GroupElementVector groupElements) {
-        return verify(
-                new MessageBlock(groupElements.map(
-                        (Function<GroupElement, GroupElementPlainText>) GroupElementPlainText::new
-                )),
-                signature,
-                verificationKey
-        );
-    }
 }